+++ src/main/java/wooteco/subway/SubwayAdminApplication.java
@@ -2,7 +2,9 @@ package wooteco.subway;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
 
+@EnableJpaAuditing
 @SpringBootApplication
 public class SubwayAdminApplication {
