+++ src/main/java/wooteco/subway/domain/station/Station.java
@@ -1,23 +1,12 @@
 package wooteco.subway.domain.station;
 
-import java.time.LocalDateTime;
-import java.util.Objects;
-
 import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
 
-import org.springframework.data.annotation.CreatedDate;
+import wooteco.subway.domain.BaseEntity;
 
 @Entity
-public class Station {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+public class Station extends BaseEntity {
     private String name;
-    @CreatedDate
-    private LocalDateTime createdAt;
 
     public Station() {
     }
@@ -35,30 +24,7 @@ public class Station {
         return this.id.equals(id);
     }
 
-    public Long getId() {
-        return id;
-    }
-
     public String getName() {
         return name;
-    }
-
-    public LocalDateTime getCreatedAt() {
-        return createdAt;
-    }
-
-    @Override
-    public boolean equals(Object o) {
-        if (this == o)
-            return true;
-        if (o == null || getClass() != o.getClass())
-            return false;
-        Station station = (Station)o;
-        return Objects.equals(id, station.id);
-    }
-
-    @Override
-    public int hashCode() {
-        return Objects.hash(id);
     }
 }
